Feature/neurological health

+++ Backend-AuraNeuro.API/Shared/Domain/Repositories/IBaseRepository.cs
@@ -4,7 +4,7 @@ public interface IBaseRepository<TEntity>
 {
     Task AddAsync(TEntity entity);
     Task<IEnumerable<TEntity>> ListAsync();
-    Task<TEntity?> FindByIdAsync(int id);
+    Task<TEntity?> FindByIdAsync(long id);
 
     void Update(TEntity entity);
